Move resources to testclasses in generics
