+++ src/test/java/spoon/test/generics/testclasses3/Bar.java
@@ -1,4 +1,4 @@
-package spoon.test.generics;
+package spoon.test.generics.testclasses3;
 
 public interface Bar<I, O> {
 	O transform(I input);
