+++ src/test/java/spoon/test/generics/testclasses3/ComparableComparatorBug.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package spoon.test.generics;
+package spoon.test.generics.testclasses3;
 
 import java.io.Serializable;
 import java.util.Comparator;
