+++ src/test/java/spoon/test/generics/testclasses3/Foo.java
@@ -1,4 +1,6 @@
-package spoon.test.generics;
+package spoon.test.generics.testclasses3;
+
+import spoon.test.generics.testclasses3.Bar;
 
 public class Foo<K, V> {
