+++ src/test/java/spoon/test/generics/testclasses3/InstanceOfMapEntryGeneric.java
@@ -1,4 +1,4 @@
-package spoon.test.generics;
+package spoon.test.generics.testclasses3;
 
 import java.util.Map;
