+++ src/test/java/spoon/test/generics/testclasses3/SimilarSignatureMethodes.java
@@ -1,4 +1,4 @@
-package spoon.test.generics;
+package spoon.test.generics.testclasses3;
 
 public class SimilarSignatureMethodes {
